+++ src/main/java/antonProject/service/RegionService.java
@@ -5,7 +5,7 @@ import antonProject.model.Region;
 import java.util.Optional;
 
 public interface RegionService {
-    Optional<Region> findByCode(Integer code);
+    Optional<Region> findByCode(Long code);
     Region save(Region region);
-    void deleteByCode(Integer code);
+    void deleteByCode(Long code);
 }
