+++ src/main/java/antonProject/shell/CarInteraction.java
@@ -16,13 +16,13 @@ public class CarInteraction {
 
     @ShellMethod(value = "Create car", key = "cc")
     public void createCar(
-            @ShellOption({"-i", "--id"}) Integer carId,
+            @ShellOption({"-i", "--id"}) Long carId,
             @ShellOption({"-c", "--cat"}) String category,
             @ShellOption({"-b", "--brand"}) String brand,
             @ShellOption({"-ns", "--numSt"}) String numberState,
             @ShellOption({"-n", "--namOwn"}) String nameOwner,
             @ShellOption({"-s", "--surOwn"}) String surnameOwner,
-            @ShellOption({"-r", "--regId"}) Integer regionId) {
+            @ShellOption({"-r", "--regId"}) Long regionId) {
         final Car car = new Car(carId, category, brand, numberState, nameOwner, surnameOwner, regionId);
         carService.save(car);
         log.debug("Машина добавлена в БД {}", car);
@@ -30,13 +30,13 @@ public class CarInteraction {
 
     @ShellMethod(value = "Update car", key = "uc")
     public void updateCar(
-            @ShellOption({"-i", "--id"}) Integer carId,
+            @ShellOption({"-i", "--id"}) Long carId,
             @ShellOption({"-c", "--cat"}) String category,
             @ShellOption({"-b", "--brand"}) String brand,
             @ShellOption({"-ns", "--numSt"}) String numberState,
             @ShellOption({"-n", "--namOwn"}) String nameOwner,
             @ShellOption({"-s", "--surOwn"}) String surnameOwner,
-            @ShellOption({"-r", "--regId"}) Integer regionId) {
+            @ShellOption({"-r", "--regId"}) Long regionId) {
         carService.findById(carId).ifPresent(car -> {
             car.setCategory(category);
             car.setBrand(brand);
@@ -51,14 +51,14 @@ public class CarInteraction {
 
     @ShellMethod(value = "Delete car", key = "dc")
     public void deleteCar(
-            @ShellOption({"-i", "--id"}) Integer carId) {
+            @ShellOption({"-i", "--id"}) Long carId) {
         carService.deleteById(carId);
         log.debug("Машина удалена из БД {}", carId);
     }
 
     @ShellMethod(value = "Find car", key = "fc")
     public void findCar(
-            @ShellOption({"-i", "--id"}) Integer carId) {
+            @ShellOption({"-i", "--id"}) Long carId) {
         carService.findById(carId).ifPresentOrElse(car -> log.info("Найдена машина {}", car), () -> log.warn("Машина {} в БД не обнаружена", carId)
         );
     }
